Moved SuccessLogger into own file.
